klar til merge, men aventer push ettersom det vil oppstå merge conflicts
