+++ PDSA_System/Server/Controllers/AuthController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using PDSA_System.Shared.Models;
 using PDSA_System.Server.Models;
 using Dapper;
