+++ PDSA_System/Server/Controllers/TeamController.cs
@@ -1,10 +1,11 @@
 using Microsoft.AspNetCore.Mvc;
-using PDSA_System.Server.Models;
+using PDSA_System.Shared.Models;
 using Dapper;
 
 namespace PDSA_System.Server.Controllers
 {
     [Route("/api/[controller]")]
+    [ApiController]
     public class TeamController : Controller
     {
         private readonly IConfiguration _configuration;
@@ -15,7 +16,7 @@ namespace PDSA_System.Server.Controllers
         }
 
         //Dette er en GET request som henter ut alle lagene i databasen.
-        [HttpGet]
+        [HttpGet("/api/[controller]/")]
         public async Task<ActionResult<List<Team>>> GetAllTeams()
         {
             // denne linjen henter ut connectionstringen fra appsettings.json
@@ -48,30 +49,30 @@ namespace PDSA_System.Server.Controllers
         Denne Metoden lager en Tuppel/Team i DB.
         */
 
-        [HttpPost]
+        [HttpPost("/api/[controller]/OpprettTeam/")]
         // denne linjen sier at denne metoden skal kjøres når det kommer en POST request
         public async Task<ActionResult<List<Team>>> CreateTeam(Team team)
         {
             var connString = _configuration.GetValue<string>("ConnectionStrings:DefaultConnection");
             using var conn = new DbHelper(connString).Connection;
-            await conn.ExecuteAsync("INSERT INTO Team (Navn, AvdelingId) VALUES (@Navn, @AvdelingId)", team);
+            await conn.ExecuteAsync("INSERT INTO Team(TeamLederId, Navn, AvdelingId) VALUES (@TeamLederId, @Navn, @AvdelingId)", team);
 
             // Await conn.ExecudeAsync betyr at vi venter på at denne linjen skal bli ferdig før vi fortsetter med neste linje.
 
-            return Ok(team); // denne linjen returnerer statuskode 200 og teamet som ble lagt til i database
+            return Ok(); // denne linjen returnerer statuskode 200 og teamet som ble lagt til i database
         }
 
 
         /* EditTeam
         * Denne metoden oppdaterer et Team.
         */
-        [HttpPut]
+        [HttpPut("/api/[controller]/updateteam/{TeamId}")]
         public async Task<ActionResult<List<Team>>> EditTeam(Team team)
         {
             var connString = _configuration.GetValue<string>("ConnectionStrings:DefaultConnection");
             using var conn = new DbHelper(connString).Connection;
 
-            await conn.ExecuteAsync("UPDATE Team SET Navn = @Navn, AvdelingId = @AvdelingId WHERE TeamId = @TeamId",
+            await conn.ExecuteAsync("UPDATE Team SET TeamLederId = @TeamLederId, Navn = @Navn, AvdelingId = @AvdelingId WHERE TeamId = @TeamId",
                 team);
