+++ PDSA_System/Server/appsettings.json
@@ -10,8 +10,9 @@
         "Secret": "ft2NueB9H8N9p8RpVvWfwW_CHANGE_ME"
     },
     "ConnectionStrings": {
-        "DefaultConnection": "server=localhost;user=root;database=NordicDoor;port=3306;password=PASSORD;"
+        "DefaultConnection": "server=localhost;user=root;database=NordicDoor;port=3306;password=vrolf; Allow User Variables=True;"
     }
-}
+    }
+
