+++ PDSA_System/Shared/Models/ByttPassord.cs
@@ -1,0 +1,8 @@
+namespace PDSA_System.Shared.Models;
+
+public class ByttPassord
+{
+    public string? Passord { get; set; }
+    public int BrukerId { get; set; }
+}
+
