+++ PDSA_System/Shared/Models/LoginFormData.cs
@@ -1,4 +1,4 @@
-namespace PDSA_System.Server.Models;
+namespace PDSA_System.Shared.Models;
 
 public class LoginFormData
 {
