+++ PDSA_System/Shared/Models/Team.cs
@@ -1,0 +1,15 @@
+
+namespace PDSA_System.Shared.Models;
+
+    public class Team
+    {
+     public int TeamId { get; set; }
+
+    public int TeamLederId { get; set; } 
+
+    public string? Navn { get; set; }
+
+    public int AvdelingId { get; set; }
+
+    }
+
